requirements.txt
PyTMX==3.32
jsmin==3.0.1
pathfinding==1.0.10
pygame-ce==2.5.0
